♻️ Improvements to adding/changing participant
- Trim name when adding
- Quicker response times
- Delete unused code
