--- apps/web/src/utils/string-to-value.ts
@@ -1,7 +1,0 @@
-export const stringToValue = (seed: string): number => {
-  let num = 0;
-  for (let i = 0; i < seed.length; i++) {
-    num += seed.charCodeAt(i);
-  }
-  return num;
-};
