+++ packages/backend/trpc/routers/polls/participants.ts
@@ -1,5 +1,6 @@
 import { prisma } from "@rallly/database";
 import { TRPCError } from "@trpc/server";
+import { waitUntil } from "@vercel/functions";
 import { z } from "zod";
 
 import { createToken } from "../../../session";
@@ -158,7 +159,7 @@ export const participants = router({
         );
       }
 
-      await Promise.all(emailsToSend);
+      waitUntil(Promise.all(emailsToSend));
 
       return participant;
     }),
